+++ functions/package.json
@@ -2,34 +2,25 @@
   "name": "functions",
   "description": "Cloud Functions for Firebase",
   "scripts": {
-    "lint": "eslint --ext .js,.ts .",
-    "build": "tsc",
-    "build:watch": "tsc --watch",
-    "serve": "npm run build && firebase emulators:start --only functions",
-    "shell": "npm run build && firebase functions:shell",
+    "lint": "eslint .",
+    "serve": "firebase emulators:start --only functions",
+    "shell": "firebase functions:shell",
     "start": "npm run shell",
-    "deploy": "tsc && firebase deploy --only functions:api",
+    "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "16"
+    "node": "14"
   },
-  "main": "lib/index.js",
+  "main": "index.js",
   "dependencies": {
-    "axios": "^1.3.6",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "firebase-admin": "^11.7.0",
-    "firebase-functions": "^4.2.0"
+    "firebase-admin": "^10.0.2",
+    "firebase-functions": "^3.14.1"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.12.0",
-    "@typescript-eslint/parser": "^5.12.0",
-    "eslint": "^8.9.0",
+    "eslint": "^7.6.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-import": "^2.25.4",
-    "firebase-functions-test": "^3.0.0",
-    "typescript": "^4.9.0"
+    "firebase-functions-test": "^0.2.0"
   },
   "private": true
 }
