--- tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "CommonJS",
-    "noImplicitReturns": true,
-    "moduleResolution": "node",
-    "noUnusedLocals": true,
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2017"
-  },
-  "compileOnSave": true,
-  "include": ["src", "tsconfig.json"],
-  "typeRoots": ["@types"]
-}
